+++ NEWS.md
@@ -1,4 +1,4 @@
-# TreeTools 1.4.5.9007 (development)
+# TreeTools 1.4.5.9008 (development)
 
 
 ## New functions